Implement optimization logic This version now needs to be invoked as: ./BeeieOptimizer <path to ROM> <path to Painting64 binary> <path to Painting64 config>
After build, the contents of `place in bin directory` must be copied to the location of the executable.

It optimizes the collision by extracting the collision mesh, and running over it with the Blender 4 script blendmaballs.py. Blender 4 is also required.

The visible geometry is manually optimized:
- texture data deduplication is performed
- there is some attempt at removing repetitive commands

My test ROM goes from 74.82 MiB to 70.03 MiB, and collision performance is indeed improved, mileage may vary depending on the area. On real HW, which I didn't have at the time, this doesn't matter nearly as much as it does on emulator.

Moving textures aren't resolved, this might be what causes Chomp's Cloudy Isles to hang on load, and other things may also be broken trol